+++ tests/test_non_default_metrics.py
@@ -130,7 +130,7 @@ def assert_request_count(
 
 def test_http_request_content_length_bytes_with_handler():
     app = create_app()
-    Instrumentator().add(metrics.http_request_content_length_bytes()).instrument(
+    Instrumentator(excluded_handlers=["/metrics"]).add(metrics.http_request_content_length_bytes()).instrument(
         app
     ).expose(app)
     client = TestClient(app)
@@ -145,7 +145,7 @@ def test_http_request_content_length_bytes_with_handler():
 
 def test_http_request_content_length_bytes_without_handler():
     app = create_app()
-    Instrumentator().add(
+    Instrumentator(excluded_handlers=["/metrics"]).add(
         metrics.http_request_content_length_bytes(should_drop_handler=True)
     ).instrument(app).expose(app)
     client = TestClient(app)
@@ -164,7 +164,7 @@ def test_http_request_content_length_bytes_without_handler():
 
 def test_http_request_content_length_bytes_no_cl():
     app = create_app()
-    Instrumentator().add(metrics.http_request_content_length_bytes()).instrument(
+    Instrumentator(excluded_handlers=["/metrics"]).add(metrics.http_request_content_length_bytes()).instrument(
         app
     ).expose(app)
     client = TestClient(app)
@@ -183,7 +183,7 @@ def test_http_request_content_length_bytes_no_cl():
 
 def test_http_response_content_length_bytes_with_handler():
     app = create_app()
-    Instrumentator().add(metrics.http_response_content_length_bytes()).instrument(
+    Instrumentator(excluded_handlers=["/metrics"]).add(metrics.http_response_content_length_bytes()).instrument(
         app
     ).expose(app)
     client = TestClient(app)
@@ -213,7 +213,7 @@ def test_http_response_content_length_bytes_with_handler():
 
 def test_http_response_content_length_bytes_without_handler():
     app = create_app()
-    Instrumentator().add(
+    Instrumentator(excluded_handlers=["/metrics"]).add(
         metrics.http_response_content_length_bytes(should_drop_handler=True)
     ).instrument(app).expose(app)
     client = TestClient(app)
@@ -246,7 +246,7 @@ def test_http_response_content_length_bytes_without_handler():
 
 def test_http_content_length_bytes_with_handler_and_data():
     app = create_app()
-    Instrumentator().add(metrics.http_content_length_bytes()).instrument(app).expose(app)
+    Instrumentator(excluded_handlers=["/metrics"]).add(metrics.http_content_length_bytes()).instrument(app).expose(app)
     client = TestClient(app)
 
     client.get("/", data="some data")
@@ -266,7 +266,7 @@ def test_http_content_length_bytes_with_handler_and_data():
 
 def test_http_content_length_bytes_with_handler_no_data():
     app = create_app()
-    Instrumentator().add(metrics.http_content_length_bytes()).instrument(app).expose(app)
+    Instrumentator(excluded_handlers=["/metrics"]).add(metrics.http_content_length_bytes()).instrument(app).expose(app)
     client = TestClient(app)
 
     client.get("/")
@@ -286,7 +286,7 @@ def test_http_content_length_bytes_with_handler_no_data():
 
 def test_http_content_length_bytes_without_handler():
     app = create_app()
-    Instrumentator().add(
+    Instrumentator(excluded_handlers=["/metrics"]).add(
         metrics.http_content_length_bytes(should_drop_handler=True)
     ).instrument(app).expose(app)
     client = TestClient(app)